+++ tests/test_non_regression_fpdu.c
@@ -451,20 +451,20 @@ static int test_decap_fpdus(const bool ignore_malformed, const char *const src_f
 			u_int8_t frag_id;
 			for (frag_id = 0; frag_id < 8; ++frag_id) {
 				printf("===\tFrag ID %u\n", frag_id);
-				printf("===\treceiver received:          %zu\n",
-				       rle_receiver_stats_get_counter_sdus_received(receiver, frag_id));
-				printf("===\treceiver reassembled:       %zu\n",
-				       rle_receiver_stats_get_counter_sdus_reassembled(receiver, frag_id));
-				printf("===\treceiver lost:              %zu\n",
-				       rle_receiver_stats_get_counter_sdus_lost(receiver, frag_id));
-				printf("===\treceiver dropped:           %zu\n",
-				       rle_receiver_stats_get_counter_sdus_dropped(receiver, frag_id));
-				printf("===\treceiver bytes received:    %zu\n",
-				       rle_receiver_stats_get_counter_bytes_received(receiver, frag_id));
-				printf("===\treceiver bytes reassembled: %zu\n",
-				       rle_receiver_stats_get_counter_bytes_reassembled(receiver, frag_id));
-				printf("===\treceiver bytes dropped:     %zu\n",
-				       rle_receiver_stats_get_counter_bytes_dropped(receiver, frag_id));
+				printf("===\treceiver received:          %llu\n",
+				       (long long unsigned int)rle_receiver_stats_get_counter_sdus_received(receiver, frag_id));
+				printf("===\treceiver reassembled:       %llu\n",
+						(long long unsigned int)rle_receiver_stats_get_counter_sdus_reassembled(receiver, frag_id));
+				printf("===\treceiver lost:              %llu\n",
+						(long long unsigned int)rle_receiver_stats_get_counter_sdus_lost(receiver, frag_id));
+				printf("===\treceiver dropped:           %llu\n",
+						(long long unsigned int)rle_receiver_stats_get_counter_sdus_dropped(receiver, frag_id));
+				printf("===\treceiver bytes received:    %llu\n",
+						(long long unsigned int)rle_receiver_stats_get_counter_bytes_received(receiver, frag_id));
+				printf("===\treceiver bytes reassembled: %llu\n",
+						(long long unsigned int)rle_receiver_stats_get_counter_bytes_reassembled(receiver, frag_id));
+				printf("===\treceiver bytes dropped:     %llu\n",
+						(long long unsigned int)rle_receiver_stats_get_counter_bytes_dropped(receiver, frag_id));
 				printf("\n");
 			}
 		}
